Fix reported memory usage on Linux
The os-utils module reports erroneous used memory percentage, in
particular, it considers memory used by the kernel's disk caching
subsystem as used memory, when it should in reality be considered free
as it is transparently let go when memory is low. See
http://linuxatemyram.com.

This changeset fixes this by detecting if we're on Linux, and if so,
runs `free -m` in a child process and screenscrapes its output. os-utils
does provide a .freeCommand() function which is supposed to be doing
this, but it also returns erroneous numbers.