[solarized] Add initial solarized dark theme
I've noticed that there's not a solarized dark theme for vtop. It looks
a bit crap without it when people reference this repository in various
popular dotfile repositories. Lets add a basic theme that corrects that.

Signed-off-by: Chris M <millscj01@gmail.com>

We should probably use a lighter border box so it matches with the
{white-fg}{/white-fg} tags that're hardcoded around headerText and
percent